+++ hal_drv/src/hal_drv_flash.c
@@ -4,18 +4,18 @@
 
 
 
-uint8_t hal_drv_flash_write(uint32_t addr, void *data, size_t len)
+int hal_drv_flash_write(uint32_t addr, void *data, size_t len)
 {
     return ql_embed_nor_flash_write(addr, data, len);
 }
 
 
-uint8_t hal_drv_flash_read(uint32_t addr, void *data, size_t len)
+int hal_drv_flash_read(uint32_t addr, void *data, size_t len)
 {
     return ql_embed_nor_flash_read(addr, data, len);
 }
 
-uint8_t hal_drv_flash_erase(uint32_t addr, size_t len)
+int hal_drv_flash_erase(uint32_t addr, size_t len)
 {
     return ql_embed_nor_flash_erase(addr, len);
 }
